server/requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.5
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-environ==0.12.0
django-js-asset==3.1.2
django-modeltranslation==0.19.16
django-unfold==0.64.2
djangorestframework==3.16.1
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.0
sqlparse==0.5.3
uritemplate==4.2.0